Lots of cleanup to split up the config file. Add avante.
