+++ lua/plugins/avante.lua
@@ -1,0 +1,80 @@
+-- AI code completion similar to Cursor
+return {
+  'yetone/avante.nvim',
+  event = 'VeryLazy',
+  version = false, -- Never set this value to "*"! Never!
+  opts = {
+    -- NOTE: For most providers you must also provide an API key as an env var
+    -- claude: ANTHROPIC_API_KEY
+    -- openai: OPENAI_API_KEY
+    -- azure openai: AZURE_OPENAI_KEY
+    -- amazon bedrock: BEDROCK_KEYS
+    -- aihubmix: AIHUBMIX_API_KEY
+
+    -- provider = "ollama",
+    -- vendors = {
+    --   ollama = {
+    --     __inherited_from = "openai",
+    --     api_key_name = "",
+    --     endpoint = "http://127.0.0.1:11434/v1",
+    --     model = "deepseek-r1:8b",
+    --   },
+    -- },
+
+    provider = 'claude',
+    claude = {
+      model = 'claude-3-7-sonnet-latest',
+    },
+
+    -- provider = 'openai',
+    -- openai = {
+    --   endpoint = 'https://api.openai.com/v1',
+    --   model = 'gpt-4o', -- your desired model (or use gpt-4o, etc.)
+    --   timeout = 30000, -- Timeout in milliseconds, increase this for reasoning models
+    --   temperature = 0,
+    --   max_completion_tokens = 8192, -- Increase this to include reasoning tokens (for reasoning models)
+    --   --reasoning_effort = "medium", -- low|medium|high, only used for reasoning models
+    -- },
+  },
+  -- if you want to build from source then do `make BUILD_FROM_SOURCE=true`
+  build = 'make',
+  -- build = "powershell -ExecutionPolicy Bypass -File Build.ps1 -BuildFromSource false" -- for windows
+  dependencies = {
+    'nvim-treesitter/nvim-treesitter',
+    'stevearc/dressing.nvim',
+    'nvim-lua/plenary.nvim',
+    'MunifTanjim/nui.nvim',
+    --- The below dependencies are optional,
+    'echasnovski/mini.pick', -- for file_selector provider mini.pick
+    'nvim-telescope/telescope.nvim', -- for file_selector provider telescope
+    'hrsh7th/nvim-cmp', -- autocompletion for avante commands and mentions
+    'ibhagwan/fzf-lua', -- for file_selector provider fzf
+    'nvim-tree/nvim-web-devicons', -- or echasnovski/mini.icons
+    'zbirenbaum/copilot.lua', -- for providers='copilot'
+    {
+      -- support for image pasting
+      'HakonHarnes/img-clip.nvim',
+      event = 'VeryLazy',
+      opts = {
+        -- recommended settings
+        default = {
+          embed_image_as_base64 = false,
+          prompt_for_file_name = false,
+          drag_and_drop = {
+            insert_mode = true,
+          },
+          -- required for Windows users
+          use_absolute_path = true,
+        },
+      },
+    },
+    {
+      -- Make sure to set this up properly if you have lazy=true
+      'MeanderingProgrammer/render-markdown.nvim',
+      opts = {
+        file_types = { 'markdown', 'Avante' },
+      },
+      ft = { 'markdown', 'Avante' },
+    },
+  },
+}
