+++ lua/plugins/goto-preview.lua
@@ -1,0 +1,7 @@
+-- Provide a preview of the goto
+return {
+  'rmagatti/goto-preview',
+  dependencies = { 'rmagatti/logger.nvim' },
+  event = 'BufEnter',
+  config = true, -- necessary as per https://github.com/rmagatti/goto-preview/issues/88
+}
