+++ lua/plugins/typescript-tools.lua
@@ -1,0 +1,9 @@
+return {
+  'pmizio/typescript-tools.nvim',
+  dependencies = { 'nvim-lua/plenary.nvim', 'neovim/nvim-lspconfig' },
+  opts = {
+    baseIndentSize = 4,
+    indentSive = 4,
+    tabSize = 4,
+  },
+}
